Fix issue with environment variable coming up nil
The module must be getting initialized before the environment
variable is available. Putting the retrieval of the environment
variable at the point where the request is made delays it. In any
case, it would be better to initialize the module with the IP of
k8s explicitly, but this gets it working enough to test.