Add bare bones ability to run a Concourse task using `fly`
The command line client for Concourse is called `fly`. It has the
ability to execute a single task using the `fly execute` command.
These additions to the HTTP routes makes it possible to kick off
a Kubernetes job by running that command with a task input file.
There is not yet any proper handling of log messages, and the
response is not what the `fly` executable expects, but it
nonetheless does start the job on Kubernetes.