Get one step further on getting `fly execute` to work
Before `fly execute` can upload an input directory for a task, it
creates a "pipe" for reading and writing. This adds the REST API
calls for creating a pipe and returning a response that tells
`fly` the URL to check for events. Nothing is functional at this
point, but the fly command will exit 0.